perf: use kitex BinaryProtocol replace apache BinaryProtocol for buffered protocol
