+++ pkg/utils/thrift.go
@@ -48,7 +48,7 @@ func (t *ThriftMessageCodec) Encode(method string, msgType athrift.TMessageType,
 	_ = thrift.Binary.WriteMessageBegin(b, method, thrift.TMessageType(msgType), seqID)
 	buf := &bytes.Buffer{}
 	buf.Write(b)
-	if err := apache.ThriftWrite(apache.NewBufferTransport(buf), msg); err != nil {
+	if err := apache.ThriftWrite(buf, msg); err != nil {
 		return nil, err
 	}
 	return buf.Bytes(), nil
@@ -71,7 +71,7 @@ func (t *ThriftMessageCodec) Decode(b []byte, msg athrift.TStruct) (method strin
 		}
 		return
 	}
-	err = apache.ThriftRead(apache.NewBufferTransport(bytes.NewBuffer(b)), msg)
+	err = apache.ThriftRead(bytes.NewBuffer(b), msg)
 	return
 }
 
@@ -79,7 +79,7 @@ func (t *ThriftMessageCodec) Decode(b []byte, msg athrift.TStruct) (method strin
 // Notice: Binary generic use Encode instead of Serialize.
 func (t *ThriftMessageCodec) Serialize(msg athrift.TStruct) ([]byte, error) {
 	buf := &bytes.Buffer{}
-	if err := apache.ThriftWrite(apache.NewBufferTransport(buf), msg); err != nil {
+	if err := apache.ThriftWrite(buf, msg); err != nil {
 		return nil, err
 	}
 	return buf.Bytes(), nil
@@ -89,7 +89,7 @@ func (t *ThriftMessageCodec) Serialize(msg athrift.TStruct) ([]byte, error) {
 // Notice: Binary generic use Decode instead of Deserialize.
 func (t *ThriftMessageCodec) Deserialize(msg athrift.TStruct, b []byte) (err error) {
 	buf := bytes.NewBuffer(b)
-	return apache.ThriftRead(apache.NewBufferTransport(buf), msg)
+	return apache.ThriftRead(buf, msg)
 }
 
 // MarshalError convert go error to thrift exception, and encode exception over buffered binary transport.
